src/services/auth/requirements.txt
grpcio
grpcio-tools
asyncpg
pydantic
python-jose
passlib